Fix die spin animation lock
Due to the double setTimout on the D20 Spin Animation it could reset in the middle of a click. Clicking on the die right after the first timout ended (once the text showed up) would stack the call for the second animation in the stack, effectively overriding the "click spin" animation after 0,5s.
Removed the logic from a second setTimout and added a delay for the CSS animation instead.